backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
requests
qdrant-client
openai
livekit-api
python-multipart
